some minor error resolved
the file name could not be uppercase. this was generating a tree error whether or not if i create a new file with the same name it will generate the same error so i had to make a new file with differnet name. previously the file name was homepage.dart and now it is home.dart and it seems to resolve the issue.